+++ Vec2.java
@@ -1,16 +1,16 @@
 public class Vec2
 {
-   private int x;
-   private int y;
+   private double x;
+   private double y;
    
-   public Vec2( int x,int y ) 
+   public Vec2( double x,double y ) 
    {
       this.x = x;
       this.y = y;
    }
    
-   public int getX(){ return this.x; }
-   public int getY(){ return this.y; }
+   public double getX(){ return this.x; }
+   public double getY(){ return this.y; }
    
    public static void displayVector( Vec2 v ) 
    {
@@ -22,5 +22,5 @@ public class Vec2
    {
       Vec2 result = new Vec2( this.x + rhs.x,this.y + rhs.y );
       return result;
-   }
+   } 
 }
